+++ pgmax/infer/bp_state.py
